pyproject.toml
[tool.poetry]
name = "mixician"
version = "0.0.4"
description = "An advanced hybrid ranking engine for recommendation systems, designed to automate the optimization of algorithms and parameters tailored to diverse business objectives."
authors = ["Yin Cheng <yin.sjtu@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yinsn/Mixician"
repository = "https://github.com/yinsn/Mixician"
packages = [{ include = "mixician", from = "src" }]

[tool.poetry.dependencies]
jax = "^0.4.24"
matplotlib = "^3.8.3"
numpyro = "^0.13.2"
optuna = "^3.5.0"
pandas = "^2.2.0"
pydantic = "^2.6.1"
python = "^3.9"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
